modemmanager: Remove band-setting implementation to improve connections
The $NWBAND command seems to disturb the internal state of the modem
such that it is unlikely to connect to the network, and produces
widely varying error codes. We don't actively use this functionality,
it's just that setting "use all bands" is part of the normal
ModemManager simple-connect sequence. Remove it so it doesn't get
triggered.

BUG=chromium-os:30064
TEST=repeatedly connect and disconnect

Change-Id: I8b98a92e3c070e025f66deb2af43824fc5c6e59f