bearer: notify final disconnect status using mm_bearer_report_connection_status
Some MMBearer methods to disconnect the bearer assume that the bearer is
disconnected even if the operation fails in a subclass. This leaves MMBearer
and MMBroadbandBearer states out of sync.

This patch notifies the bearer object of the final disconnection using
mm_bearer_report_connection_status. This gives the subclasses a chance to update
their state even when the disconnection attempt fails.

BUG=chromium:311660
TEST=(1) Connect to cellular network.
     (2) Attempt to disconnect such that disconnect fails. (One way to do this
         is to move out of coverage area). Verify no error from logs.
     (3) Move device back to coverage area. Verify that auto-connect happens
         properly.

Change-Id: Iab27264d448e6272fae4df5b88acad7c5f1bdbaf
Reviewed-on: https://chromium-review.googlesource.com/174797
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>