port-probe: Increase AT probe command timeout to 7 seconds for altair.
This patch increases the response timeout for the probe AT commands for
altair modems.

We've been noticing some altair modems taking upto 6 seconds to respond to the
initial probe command after a reset which results in modem-manager
timing out and sending a second probe command. The modem sends a response
after about 6 seconds for the initial probe command which modem-manager
treats as response to second probe command and this results in the
modem-manager and modem going out of sync because the modem's second
probe response is treated as response to the next initialization AT command
sent by modem-manager and so on.

BUG=chrome-os-partner:33718
TEST='test_that ${host} network_3GModemPresent'
Change-Id: Iad8b0786327b153fd95c8ee4516f352325a42cf7
Reviewed-on: https://chromium-review.googlesource.com/239960
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>