udev: activate ModemManager.service upon seeing a candidate
Let's bring ModemManager up when we see a possible modem. This allows us
to avoid enabling ModemManager unconditionally, which is good on
machines that don't have modems and do the right thing when a modem is
attached afterwards.

This works best if we're not too zealous about marking devices
ID_MM_CANDIDATE.