net-port-mapper: store information of created net interfaces
Add a class to hold the information of net interfaces created inside
Modem Manager. This class is used to link the net interface and the
control port that was used to create it. Some new SoCs (i.e. Qrtr based)
allow the creation of net interfaces by userspace(i.e. ModemManager) and
these net interfaces show up as virtual net interfaces with incomplete
sysfs information. This class allows the probing and matching of those
interfaces with partial sysfs information.