CHROMIUM: broadband-modem-qmi: do not override modem-originated profiles
For the LTE attach PDN, Qualcomm advises to not modify the pre-existing
profiles in the modem. According to them, this might lead to attach
issue on other networks later. They recommend to create a profile
dedicated to the application with an identifiable name.

Update the initial EPS bearer according, read the name of the current
profile and if doesn't match our constant name for ModemManager, create
a new one and later force using it with wds-lte-set-attach-pdn-list.

BUG=b:170367556
BUG=b:174818063
TEST=On PomPom, attach to the SoftBank Mobile network and manually observe
ModemManager debug traces.

Change-Id: I9eff4441f001a52edc7a0be3aa2193c79ce24819
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2659209
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>