broadband-modem-qmi: do not override modem-originated profiles
For the LTE attach PDN, Qualcomm advises to not modify the pre-existing
profiles in the modem. According to them, this might lead to attach
issue on other networks later. They recommend to create a profile
dedicated to the application with an identifiable name.

Update the initial EPS bearer according, read the name of the current
profile and if doesn't match our constant name for ModemManager, create
a new one and later force using it with wds-lte-set-attach-pdn-list.