REVERT-ME: sim-qmi: Delay sim loading by 1 second
If two sims are inserted, the FW reports IMSI of both sims to be the
same. Currently, MM uses a successful GET_CARD_STATUS as an
indication that the sim is ready. It then reads stale IMSI from the
firmware.

A workaround is to wait 1 second after GET_CARD_STATUS before reading
 IMSI. For the ideal fix, we need a method to know when the new SIM's
 IMEI has been read by the FW.

This bug does not affect us when a single sim is inserted.

BUG=b:182484170
TEST=restart modemmanager; modem status

Change-Id: I7616c2b70c038e72137ae679d327997562cc046d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2753029
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>