UPSTREAM: iface-modem: Increase unlock_required retries
If a SIM is known to exist, for e.g. if it was created during load_sim_slots,
persist a few more times before giving up on the SIM to be ready. There
are modems on which the SIM takes more than 15s to be ready, luckily,
they happen to be QMI modems where the SIM's iccid in load_sim_slots
lets us know that there is a sim.

BUG=b:185217136
TEST=none, upstream patch

Change-Id: Id6126870ce4ce708db15db7d0472018d3979e4ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2888707
Tested-by: Pavan Holla <pholla@google.com>
Auto-Submit: Pavan Holla <pholla@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>