api,port-qmi: define udev tags to configure static aggregation settings
There are Qualcomm SoCs, like the SDM845 where we don't know how to
query the WDA data settings, any type of command we use fails.

In order to allow setting up connectivity in this kind of SoCs, we'll
allow users to provide udev tags on the IPA net port that specify
which are the data aggregation settings configured.

These settings can be derived from the vendor-specific
'netmgr_config.xml' file found in Android.