CHROMIUM: broadband-modem-qmi: fix profile handling for initial EPS bearer
Follow-up to the 'Chromium'-only patch in crrev.com/c/2659209
implementing the Qualcomm recommendation for handling the profiles.

Fix the logic in the initial EPS bearer state for the case where we
power down the modem, we must then follow 2 different paths whether we
are already using our own profile or we need to create one, rather than
always doing the latter.

BUG=b:186115208
TEST=On Lazor, with the #cellular-use-attach-apn set, while already
connected to the network, go to APN UI, enter a custom APN (the same as
the prefined one) and select 'attach APN'. Check the log that we re-use
the current profile.

Change-Id: I76e39e5cfce51be53b0acc4b11fbf00f55759f4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2848501
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>