CHROMIUM: Force rmnet_data0 as the only ipa link
Now that multiplex is supported in ModemManager, the net interface for
QRTR based modems is only created when the bearer is created. Shill
doesn't know how to handle this, since it assumes that the modem will
send the net interface information as soon as the modem is created, and
not later.
This commit forces ModemManager to always use the existing rmnet_data0
interface for all bearers.

BUG=b:186213732
BUG=b:175305412
TEST=test_that $LAZOR_IP:$LAZOR_PORT suite:cellular_ota --fast
TEST=Verify internet connection in lazor when connecting and
disconnecting multiple times. Disable and enable Mobile data without
issues.

Cq-Depend: chromium:2864011, chromium:2864012
Change-Id: I2ce052ae79f7908288d1c79b6d32aa8a4e3549d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2855769
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>