CHROMIUM: broadband-modem-qmi: Use modem profile when setting empty attach APN
Since we don't always know the ip type needed for initial attach, we
will let the modem decide the initial attach settings when we don't
need a custom attach APN. This is done by selecting the first profile
in the modem, which should contain the APN that the modem negotiated
with the carrier during registration.

BUG=b:185129925
TEST=Using a tmobile SIM on lazor, set the APN to FAST.T-MOBILE.COM with
the attach option. Verify that it fails to connect because the ip type is
ipv4 only. Change the APN to LTE in the UI. Turn mobile data off and on.
Verify that the MM is getting the empty attach settings and it's choosing
the profile number 1 from the modem instead of using the CrOS_attach_PDN
profile. Change back to a custom APN and verify the MM uses the
CrOS_attach_PDN profile. Test with Verizon and AT&T SIM cards. Verified
that AT&T works with an attach APN.

Change-Id: Ibcfea7b9a6f59f0683caf980fdeb2d95b0254aef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/2898750
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>