CHROMIUM: Reland the attach APN CLs combined in one
This CL combines all the previous Attach APN CLs reverted for the
ModemManager uprev. This CL is very similar to the previous ones, except
that it uses the ProfileManager to load, delete, create and modify the
profiles.

BUG=b:200066788
BUG=b:194284044
TEST=Test on a limozeen and a lazor device with the Attach APN feature
ON. Tested with a AT&T SIM card with the APN broadband as Attach and
without attach, multiple times. Tested with a Tmobile SIM card, but test
is limited with it since tmobile doesn't use attach.

Change-Id: Id5238da898276eb1d5c9be9eded5888bf08b584c
Cq-Depend: chromium:3224806
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3233246
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Madhav . <madhavadas@google.com>