broadband-modem-qmi: store initial EPS bearer in a separate profile
When using SetInitialEpsBearerSettings on a qmi modem, we should not
reuse any of the existing profiles in the modem, since it might override
the default profile provided by the modem. According to QC, if the
default profile is overriden, the modem might no longer be able to
attach to the network unless the persistant memory of the modem is
reset.