ublox: Implement profile manager to select initial eps bearer context
When trying to set up a new pdp context (e.g. due to simple connect) on
ublox toby-L2 or MPCI-L2, the initial eps bearer context has to be selected
when it matches the APN requested and is already connected. This is done
by preferring CID 4 when searching for a PDP context to reuse.