base-modem: trigger full modem reprobe on new port notifications
If we have a valid modem object created and we receive a notification
for a new port during runtime, it may very likely mean that we created
the modem object too soon, without the full list of ports that the
device exposes.

Instead of going on with the incomplete modem object, trigger a full
reprobe right away so that we are sure we use all the available ports.

This is specially important if the port being reported is e.g. a QMI
or MBIM control port, we don't want to ignore those as it could mean
we're falling back to AT+PPP when it's not really mandatory to do so.