CHROMIUM: Toggle modem power on MBIM after attach APN
Toggle the modem power after attach APN to force the modem to reattach
to the network when the attach APN is modified. This should be handled
by the modem itself, but some modems don't do it.

BUG=b:198002531
TEST=Test with callbox on nipperkin and storo360. Verify that the
backoff timer does not affect the device after switching attach APNs.

Change-Id: Ie6af9e2923abbf3d5cef000ff2d19236c47c6e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3437278
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>