REVERT-ME: modemmanager-next: vilboz360 use antenna index 0 for sar
As per spec, setting the antenna index to 0xFFFFFFFF
should set the backoff index to all antennas on the system.
This feature is currently not supported in
Fibocom FW 18500.5001.30.05.27.12.
This will be fixed in the next MR.
Please revert this patch once the new FW lands.

BUG=b:188002987
TEST=emerge-zork modemmanager-next && cros deploy
     mmcli -m 0 --sar-set-power-level=0

Change-Id: If34d7f9409a0213f7a07e1d0d9f062e36918d8b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3570861
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Madhav <madhavadas@google.com>
Commit-Queue: Madhav <madhavadas@google.com>