CHROMIUM: Revert "broadband-modem-mbim: treat 'no eSIM profile' as initialized"
This reverts commit 9940bd9dbe88430318ae8f045001db68183b42c5.

Shill expects profileless eSIM's to leave the modem in a failed state
(This is true on all our other modems). Upstream has agreed to our
stance in principle, see
https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/commit/9940bd9dbe88430318ae8f045001db68183b42c5

However, we still need to decide on the right long term implementation.
In order to enable testing of rest of the eSIM stack, we need this revert.

BUG=b:228169333
TEST=Check that profileless eSIM's show up in a failed state

Cq-Depend: chromium:3577698
Change-Id: Ib78435585efe5f39ee07b23f527177f799b0d1f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3578812
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>