CHROMIUM: modemmanager: Workaround for reattach on fm350
This workaround is proposed to suppress the modem retry timer in case
of a PS attach failure and for initiating the next PS attach request.

This adaptation for this workaround is implemented in the plug-in
which toggles the power mode so that modem doesn't run the retry timer
and considers the new attach APN request as a fresh request.

BUG=b:225197387
TEST=Build and verify reattach manually using mmcli

Signed-off-by: Dinesh Kumar Sharma <dinesh.sharma@intel.corp-partner.google.com>
Change-Id: If2909ddc6ee23676134ccfb4ec5f594ec7d38c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3647037
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>