CHROMIUM: Remove partner profile on L850GL for AT&T 310280
There is a bug on the L850GL MR5 firmware specific to AT&T SIM cards
with the MCCMNC 310280. The bug on the modem can be contained by
removing the 'partner' attach profile. This change can be removed once a
new modem FW with the fix is provided.

BUG=b:224986971
TEST=Test with an AT&T SIM card(301280) and a regular AT&T SIM card.
Test same patch with the callbox SIM.
Verify that NL668 doesn't execute the new logic.

Change-Id: Ide1f7a74bafefdb3ea463ba4ed82a7b0b76ba0ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3761575
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>