CHROMIUM: Expand the attach APN hack for 310280 to all APNs
The initial hack was only triggered when the attach APN being used was
"broadband". This was done to reduce the scope of the hack since it was
assumed that AT&T always used "broadband" as attach APN. AT&T have just
confirmed that they use custom Attach APNs for some clients, and without
the hack they will have issues registering to the network.

BUG=b:243650576
TEST=None yet. Tested by client on canary channel.

Change-Id: Ie4cbabdf79769fbcff59f457aee279b55b6bb833
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3867759
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>