CHROMIUM: Retry SIM load operator identifier
On certain SIM cards, MM times out the first time it queries for the
'home-provider' information after the modem is reset or the device cold
boots. From tests, we see that the second query always succeeds, since
there is a ~25 seconds delay between the first and second query due to
the timeout of the first query. If reading the operator id fails, we
cannot apply specific logic to certain MCCMNCs later on.

BUG=b:244791521
TEST=Deploy to sasuke l850. Verify that when the first query times out,
MM can read the value on the retry, and the special logic for
MCCMNC==310280 gets triggered.

Change-Id: I6dd2dc70f100fd50c30bcd8b2dc825bf979600b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3871914
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Pavan Holla <pholla@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>