CHROMIUM: fibocom: ensure udev rules are run on 'bind' events
The ID_MM_PORT_IGNORE for TTYs is applied on 'bind' events, which
should be treated in the same way as the 'add', 'move' and 'change'
events.

At the same time, the rules are updated to match the exact
SUBSYSTEM (instead of the wider SUBSYSTEMS that matches the parent
devices as well), and also to use ATTRS{idVendor} instead of
ENVi{ID_VENDOR_ID} (so using the sysfs attributes instead of the env
set by the generic udev rules).

BUG=b:250667187
TEST=Run 'mmcli -m a' and ensure all the detected TTY ports are
flagged as 'ignored'.

Change-Id: I133fff7810e6b3d5add02426219d5ac567390f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/3925707
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>