UPSTREAM: port-qmi: implement the new generic 'removed' signal
Letting the MMBroadbandModemQmi listen for the notifications of the
QmiDevice was not a good idea, especially since no explicit reference
to the device object was hold in the modem object. This leads to race
conditions in which the QmiDevice outlives the MMBroadbandModemQmi and
the MMPortQmi, and when the "device-removed" signal is triggered, the
program crashes.

The MMPortQmi will now emit its own 'removed' signals when the
underlying QmiDevice emits "device-removed', ensuring that the signal
handler is properly cleared up during the MMPortQmi disposal.

(cherry picked from commit a20f2428eea3fa6fd83edeb5f1a844926b4a5b1c)

BUG=b:260590252
TEST=CQ passes

Change-Id: I7306505b5c152e2966c5a585881b9b29bc1bbcbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/4110834
Tested-by: Pavan Holla <pholla@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>