UPSTREAM: port-mbim: implement the new generic 'removed' signal
Letting the MMBroadbandModemMbim listen for the notifications of the
MbimDevice was not a good idea, especially since no explicit reference
to the device object was hold in the modem object. This leads to race
conditions in which the MbimDevice outlives the MMBroadbandModemMbim and
the MMPortMbim, and when the "device-removed" signal is triggered, the
program crashes.

The MMPortMbim will now emit its own 'removed' signals when the
underlying MbimDevice emits "device-removed', ensuring that the signal
handler is properly cleared up during the MMPortMbim disposal.

Fixes https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/issues/518

(cherry picked from commit 309a8a515bf922b679f36971bffc2999bdc6e293)

BUG=b:260590252
TEST=CQ passes

Change-Id: I3ed1acd775fe659c3467ba698d07e88f712fcadf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/modemmanager-next/+/4110835
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>
Tested-by: Pavan Holla <pholla@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>