base-modem-at: use better names for pairs of dependent cancellables
We cancel AT commands (or sequences) on cancellation of the modem's
cancellable as well as of the optional user-provided one.

However, the GTask can be associated with only one. That's okay if the
user didn't supply one -- we just use the modem's cancellable only.
Otherwise we chain the cancellables together (also cancel the user one
with the modem one) at cost of some extra complexity.

This aims to make the above a little clearer by using hopefully
better names.

Suggested-by: Aleksander Morgado <aleksandermj@chromium.org>